+++ ceammc/ext/src/lib/ceammc_fn_list.h
@@ -50,6 +50,8 @@ namespace list {
     bool calcClipIndex(int pos, size_t len, size_t* idx);
     bool calcWrapIndex(int pos, size_t len, size_t* idx);
     bool calcFoldIndex(int pos, size_t len, size_t* idx);
+
+    AtomList rotate(const AtomList& l, int steps);
 }
 }