+++ ceammc/CHANGELOG.md
@@ -43,6 +43,7 @@
     - set.symdiff
     - set.equal
 - list.seq uses only positive step. To get 6 4 2 use [list.seq 6 0 2]
+- list.rotate alias added [list.<<]
 
 ### Changed:
 - pass.if renamed to flow.pass_if (with pass_if alias)